Created a seacrh bar to check for books
Populated an array with a few book objects. Created an input element
that takes in the book's title or author's name or read status.

I wrote a for loop to iterate through the array of book objects and
display the book that matches the information provided in the seacrh
bar.